Home page canvas: Scaling canvas according to page size
We are rendering image on home page in a canvas. As, images used for rendering are curretly in FHD format, only part of the images is visible is client's widow size is small.
We are now scaling the canvas by cosidering the ration of original image size and window size.
The solution is not perfect, depending of scale factor blank horizontal or vertical lines can appear in resolution lower than FHD. Also, we need to pass split image row count from db, which right now is hard coded.